LearnJava

+++ Main.java
@@ -3,7 +3,7 @@ public class Main{
     public static void main(String[] args){
 
         int x = 123;
-        float y = 3.14f;
+        double y = 3.14f;
         System.out.println("My number is: "+y);
     }
 }
